u16 indices to decrease memory usage

+++ gpu/include/kernels/assigner.cuh
@@ -10,14 +10,11 @@ constexpr usize MAX_CLASS_LIMIT = 2048;
 namespace kernels {
 
 // TODO: this will take up a ton of memory for larger populations
-// ideas:
-// - replace usize indices with u16 (4x less space, and all indices in every problem are below 2^16 anyway)
-// - ???
 struct StudentAssignment {
     // (for class `i` in solution `j`)
     // indices of students taking this class are placed starting at
     // student_idxs[MAX_CLASS_LIMIT * (j * n_classes + i)] ...
-    thrust::device_vector<usize> students_idxs;
+    thrust::device_vector<u16> students_idxs;
     // and their count is class_counts[i * n_classes + j]
     thrust::device_vector<u32> class_counts;
     usize n_classes;
